edit workflows

+++ scripts/runner.sh
@@ -5,15 +5,19 @@ LOG="$2"
 CMD="$3"
 
 cleanup() {
+  if [[ $EUID -ne 0 ]]; then
     pkill -9 Python
     pkill -9 python
+  fi
 }
 
 runCmdWithTimeout() {
   (exec sh -c "$2") 2>>"$3" 1>/dev/null & pid=$!
   (sleep "$1" && kill -HUP $pid) 2>/dev/null & watcher=$!
   if wait $pid || ex=$? >/dev/null; then
-    pkill -HUP -P $watcher
+    if [[ $EUID -ne 0 ]]; then
+      pkill -HUP -P $watcher
+    fi
     wait $watcher
     cleanup
     echo $((ex))
